+++ python/tests/sbp/test_table.py
@@ -34,7 +34,7 @@ def test_table_count():
   Test number of available messages to deserialize.
 
   """
-  number_of_messages = 72 
+  number_of_messages = 73
   assert len(_SBP_TABLE) == number_of_messages
 
 def test_table_unqiue_count():